requirements.txt
gradio==4.27.0
python-dotenv
google-generativeai>=0.5.2
